requirements.txt
Flask
Flask-Cors
gTTS
SpeechRecognition
pydub
google-generativeai
